+++ src/components/Accounts/ChargeFeeFine/FeeFineInfo.js
@@ -24,27 +24,12 @@ class FeeFineInfo extends React.Component {
   constructor(props) {
     super(props);
     this.onChangeAmount = this.onChangeAmount.bind(this);
-    this.onChangeFeeFine = this.onChangeFeeFine.bind(this);
     this.amount = 0;
   }
 
   onChangeAmount(e) {
     this.amount = parseFloat(e.target.value || 0).toFixed(2);
-    this.props.onChangeFeeFine(this.amount || 0);
-  }
-
-  onChangeFeeFine(e) {
-    const {
-      form: { change },
-      feefines,
-    } = this.props;
-    const feeFineId = e.target.value;
-    const feefine = feefines.find(f => f.id === feeFineId) || {};
-    this.amount = feefine.defaultAmount || 0;
-    this.amount = parseFloat(this.amount).toFixed(2);
-    const defaultAmount = parseFloat(feefine.defaultAmount || 0).toFixed(2);
-    this.props.onChangeFeeFine(defaultAmount, feeFineId);
-    change('amount', defaultAmount);
+    this.props.onChangeFeeFine(this.amount);
   }
 
   onBlurAmount = () => {
@@ -58,6 +43,7 @@ class FeeFineInfo extends React.Component {
       ownerOptions,
       feefineList,
       onChangeOwner,
+      onChangeFeeFine,
     } = this.props;
 
     return (
@@ -111,7 +97,7 @@ class FeeFineInfo extends React.Component {
                           disabled={isPending.feefines}
                           dataOptions={feefineList}
                           placeholder={placeholder}
-                          onChange={this.onChangeFeeFine}
+                          onChange={onChangeFeeFine}
                         />
                       )}
                     </FormattedMessage>