+++ src/components/Accounts/accountFunctions.js
@@ -1,4 +1,7 @@
-import { paymentStatusesAllowedToRefund } from '../../constants';
+import {
+  paymentStatusesAllowedToRefund,
+  waiveStatuses,
+} from '../../constants';
 
 export function count(array) {
   const list = [];
@@ -41,11 +44,35 @@ export function handleFilterClear(name) {
   return state;
 }
 
-export function calculateSelectedAmount(accounts, isRefundAction = false) {
+function getPaidActions(feeFineActions = []) {
+  return feeFineActions.filter(({ typeAction }) => paymentStatusesAllowedToRefund.includes(typeAction));
+}
+
+function getWaiveActions(feeFineActions = []) {
+  return feeFineActions.filter(({ typeAction }) => waiveStatuses.includes(typeAction));
+}
+
+export function calculateSelectedAmount(accounts, isRefundAction = false, feeFineActions = []) {
   const selected = accounts.reduce((s, { amount, remaining }) => {
-    return isRefundAction
-      ? s + parseFloat((amount - remaining) * 100)
-      : s + parseFloat(remaining * 100);
+    if (isRefundAction) {
+      const waiveActions = getWaiveActions(feeFineActions);
+      const waivedAmount = waiveActions.reduce((a, { amountAction }) => {
+        return a + parseFloat(amountAction * 100);
+      }, 0);
+
+      return s + parseFloat((amount - remaining) * 100) - waivedAmount;
+    } else {
+      return s + parseFloat(remaining * 100);
+    }
+  }, 0);
+
+  return parseFloat(selected / 100).toFixed(2);
+}
+
+export function calculateRefundSelectedAmount(feeFineActions) {
+  const paidActions = getPaidActions(feeFineActions);
+  const selected = paidActions.reduce((s, { amountAction }) => {
+    return s + parseFloat((amountAction) * 100);
   }, 0);
 
   return parseFloat(selected / 100).toFixed(2);
@@ -94,9 +121,13 @@ export function accountRefundInfo(account) {
   return { hasBeenPaid, paidAmount };
 }
 
-export function isRefundAllowed(account) {
-  const { hasBeenPaid, paidAmount } = accountRefundInfo(account);
-  return hasBeenPaid && paidAmount > 0;
+// export function isRefundAllowed(account) {
+//   const { hasBeenPaid, paidAmount } = accountRefundInfo(account);
+//   return hasBeenPaid && paidAmount > 0;
+// }
+
+export function isRefundAllowed() {
+  return false;
 }
 
 export function calculateTotalPaymentAmount(accounts = []) {