+++ src/constants.js
@@ -75,3 +75,36 @@ export const paymentStatusesAllowedToRefund = [
   'Transferred fully',
   'Transferred partially',
 ];
+
+export const waiveStatuses = [
+  'Waived fully',
+  'Waived partially'
+];
+
+export const reportColumns = [
+  'borrower.name',
+  'borrower.barcode',
+  'borrowerId',
+  'dueDate',
+  'loanDate',
+  'loanPolicy.name',
+  'loanPolicyId',
+  'loanId',
+  'feeFine',
+  'item.title',
+  'item.materialType.name',
+  'item.status.name',
+  'item.barcode',
+  'item.callNumberComponents.prefix',
+  'item.callNumberComponents.callNumber',
+  'item.callNumberComponents.suffix',
+  'item.volume',
+  'item.enumeration',
+  'item.chronology',
+  'item.copyNumber',
+  'item.contributors',
+  'item.location.name',
+  'item.instanceId',
+  'item.holdingsRecordId',
+  'itemId'
+];