+++ test/bigtest/interactors/users.js
@@ -7,11 +7,22 @@ import {
   isPresent,
   isVisible,
   scoped,
+  selectable,
 } from '@bigtest/interactor';
+
+import MultiSelectInteractor from '@folio/stripes-components/lib/MultiSelection/tests/interactor';
 
 @interactor class ActiveUserCheckbox {
   clickActive = clickable('#clickable-filter-active-active');
   clickInactive= clickable('#clickable-filter-active-inactive');
+}
+
+@interactor class DepartmentsFilter {
+  multiSelect = new MultiSelectInteractor('#departments-filter');
+  isAccordionPresent = isPresent('#users-filter-accordion-departments');
+  whenLoaded() {
+    return this.when(() => isPresent('#departments-filter'));
+  }
 }
 
 @interactor class HeaderDropdown {
@@ -21,6 +32,7 @@ import {
 @interactor class HeaderDropdownMenu {
   clickExportToCSV = clickable('#export-overdue-loan-report');
   exportBtnIsVisible = isVisible('#export-overdue-loan-report');
+  isExportBtnPresent = isPresent('#export-overdue-loan-report');
 }
 
 @interactor class SearchFieldInteractor {
@@ -31,10 +43,12 @@ export default @interactor class UsersInteractor {
   static defaultScope = '[data-test-user-instances]';
 
   activeUserCheckbox = new ActiveUserCheckbox();
+  departmentsFilter = new DepartmentsFilter();
   headerDropdownMenu = new HeaderDropdownMenu();
   searchField = new SearchFieldInteractor();
   searchButton = new Interactor('[data-test-user-search-submit]');
   searchFocused = isPresent('[data-test-user-search-input]:focus');
+  chooseSearchOption = selectable('#input-user-search-qindex');
   paneHeaderFocused = is('#users-search-results-pane [class*=paneHeader---]', ':focus');
   patronGroupsPresent = isPresent('#clickable-filter-pg-faculty');
   instancePresent = isPresent('[data-test-instance-details]');