+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change history for ui-users
 
-## 5.0.0 (IN PROGRESS)
+## 5.1.0 (IN PROGRESS)
+
+* Fix bug showing duplicated service points in add service point dialog. Fixes UIU-1892.
+* Add report icon for report menu items. Refs UIU-1505.
+* Allow search by username. Refs UIU-1707.
+* Use more efficient queries for `accounts` records. Refs UIU-1913.
+
+## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
+[Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)
+
+* Update plugins to `stripes v5`-compatible versions. Refs UIU-1901.
+
+## [5.0.0](https://github.com/folio-org/ui-users/tree/v5.0.0) (2020-10-14)
+[Full Changelog](https://github.com/folio-org/ui-users/compare/v4.0.0...v5.0.0)
 
 * modified the lost date display to include aged to lost status. Fixes UIU-1804.
 * Fixing breaking tests.
@@ -54,6 +67,12 @@
 * Correctly handle permissions modal display over edit pane. Fixes UIU-1857.
 * Adding of `refund` fee/fine single action. Refs UIU-1850.
 * Refactor actions for single list item. Refs UIU-1797.
+* Search translations of permission names. Refs UIU-1859.
+* Hide the Overdue loans report option when user doesn't have view loans permissions. Refs UIU-1858.
+* Allow renewal override if due date isn't changed. Refs UIU-1853.
+* Make the assigned service points for the user have been checked in the Add service points modal. Fixes UIU-1560.
+* Add Departments filter. Refs UIU-1355.
+* Always show the `Title` field for a loan, even if it is empty. Fixes UIU-1573.
 
 ## [4.0.0](https://github.com/folio-org/ui-users/tree/v4.0.0) (2020-06-17)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v3.0.0...v4.0.0)